makes AddEvent comp from AddNew, updates context
